* branch-0.6: (192 commits)
  DOC Update CHANGELOG
  Cmake build --parallel doesn't work as expected so falling back to passing flag via native flags
  ENH Add parallel env var back to limit threads used
  Fix CMake condition checking and argument passing
  FIX Updating commit version of rmm to match branch-0.5
  FIX Pin submodule to known working branches
  Forgot word in status message
  Forgot GTest changes
  Update CMake to pass `PARALLEL_LEVEL` environment variable to GTest and Arrow builds, pass variable through to conda build as well
  ENH Rename env var for parallel level
  ENH Remove unused env var for parallel builds
  Remove verbosity and samples from GTest
  changelog entry
  Remove relevant travis-ci references, update CMake to always pass --parallel since that's current behavior when not in Travis-CI
  Revert "Remove TravisCI disabling of parallel builds, always pass --parallel to build"
  Remove TravisCI disabling of parallel builds, always pass --parallel to build
  Update README for default ABI
  Update README for ABI default
  Update README
  DOC Add CHANGELOG section for v0.6
  ...

+++ cpp/src/io/csv/csv_reader.cu
@@ -95,8 +95,7 @@ typedef struct raw_csv_ {
     char				decimal;
     char				thousands;
 
-	// TODO: change nrows to gdf_size_type once gdf_size_type is changed to a signed integer type
-	int  				nrows;       	// number of rows of file to read. default is -1, and all rows are read in this case
+	gdf_size_type  		nrows;       	// number of rows of file to read. default is -1, and all rows are read in this case
 	gdf_size_type		skiprows;       // number of rows at the start of the file to skip, default is 0
 	gdf_size_type		skipfooter;     // number of rows at the bottom of the file to skip, default is 0
 
@@ -112,7 +111,7 @@ typedef struct column_data_ {
 	unsigned long long countInt16;
 	unsigned long long countInt32;
 	unsigned long long countInt64;
-	unsigned long long countNULL;
+	gdf_size_type countNULL;
 } column_data_t;
 
 typedef struct parsing_opts_ {
@@ -380,7 +379,7 @@ gdf_error read_csv(csv_read_arg *args)
 		auto recCount = raw_csv->num_records;
 
 		bool quotation = false;
-		for (size_t i = 1; i < raw_csv->num_records; ++i) {
+		for (gdf_size_type i = 1; i < raw_csv->num_records; ++i) {
 			if (h_uncomp_data[h_rec_starts[i] - 1] == raw_csv->quotechar) {
 				quotation = !quotation;
 				h_rec_starts[i] = raw_csv->num_bytes;
@@ -545,7 +544,7 @@ gdf_error read_csv(csv_read_arg *args)
 	if (args->input_data_form == gdf_csv_input_form::FILE_PATH)
 	{
 		close(fd);
-		munmap(map_data, raw_csv->num_bytes);
+		munmap(map_data, map_size);
 	}
 
 
@@ -553,6 +552,9 @@ gdf_error read_csv(csv_read_arg *args)
 	//--- Auto detect types of the vectors
 
 	if(args->dtype==NULL){
+		if (raw_csv->num_records == 0) {
+			checkError(GDF_INVALID_API_CALL, "read_csv: no data available for data type inference");
+		}
 
 		column_data_t *d_ColumnData,*h_ColumnData;
 
@@ -615,7 +617,7 @@ gdf_error read_csv(csv_read_arg *args)
 
 	void **d_data,**h_data;
 	gdf_valid_type **d_valid,**h_valid;
-    unsigned long long	*d_valid_count,*h_valid_count;
+    unsigned long long	*d_valid_count;
 	gdf_dtype *d_dtypes,*h_dtypes;
 
 
@@ -623,7 +625,6 @@ gdf_error read_csv(csv_read_arg *args)
 
 
 	h_dtypes 		= (gdf_dtype*)malloc (	sizeof(gdf_dtype)* (raw_csv->num_active_cols));
-	h_valid_count	= (unsigned long long*)malloc (	sizeof(unsigned long long)* (raw_csv->num_active_cols));
 	h_data 			= (void**)malloc (	sizeof(void*)* (raw_csv->num_active_cols));
 	h_valid 		= (gdf_valid_type**)malloc (	sizeof(gdf_valid_type*)* (raw_csv->num_active_cols));
 
@@ -686,45 +687,49 @@ gdf_error read_csv(csv_read_arg *args)
 	free(h_dtypes); 
 	free(h_valid); 
 	free(h_data); 
-	
-	launch_dataConvertColumns(raw_csv, d_data, d_valid, d_dtypes, d_str_cols, d_valid_count);
-	cudaDeviceSynchronize();
 
-	stringColCount=0;
-	for (int col = 0; col < raw_csv->num_active_cols; col++) {
-
-		gdf_column *gdf = cols[col];
-
-		if (gdf->dtype != gdf_dtype::GDF_STRING)
-			continue;
-
-		NVStrings* const stringCol = NVStrings::create_from_index(h_str_cols[stringColCount],size_t(raw_csv->num_records));
-		if ((raw_csv->quotechar != '\0') && (raw_csv->doublequote==true)) {
-			// In PANDAS, default of enabling doublequote for two consecutive
-			// quotechar in quote fields results in reduction to single
-			std::string quotechar = std::string(&raw_csv->quotechar);
-			std::string doublequotechar = quotechar + raw_csv->quotechar;
-			gdf->data = stringCol->replace(doublequotechar.c_str(), quotechar.c_str());
-			NVStrings::destroy(stringCol);
+	if (raw_csv->num_records != 0) {
+		error = launch_dataConvertColumns(raw_csv, d_data, d_valid, d_dtypes, d_str_cols, d_valid_count);
+		if (error != GDF_SUCCESS) {
+			return error;
 		}
-		else {
-			gdf->data = stringCol;
+		// Sync with the default stream, just in case create_from_index() is asynchronous 
+		cudaStreamSynchronize(0);
+
+		stringColCount=0;
+		for (int col = 0; col < raw_csv->num_active_cols; col++) {
+
+			gdf_column *gdf = cols[col];
+
+			if (gdf->dtype != gdf_dtype::GDF_STRING)
+				continue;
+
+			NVStrings* const stringCol = NVStrings::create_from_index(h_str_cols[stringColCount],size_t(raw_csv->num_records));
+			if ((raw_csv->quotechar != '\0') && (raw_csv->doublequote==true)) {
+				// In PANDAS, default of enabling doublequote for two consecutive
+				// quotechar in quote fields results in reduction to single
+				const string quotechar(1, raw_csv->quotechar);
+				const string doublequotechar(2, raw_csv->quotechar);
+				gdf->data = stringCol->replace(doublequotechar.c_str(), quotechar.c_str());
+				NVStrings::destroy(stringCol);
+			}
+			else {
+				gdf->data = stringCol;
+			}
+
+			RMM_TRY( RMM_FREE( h_str_cols [stringColCount], 0 ) );
+
+			stringColCount++;
 		}
 
-		RMM_TRY( RMM_FREE( h_str_cols [stringColCount], 0 ) );
+		vector<unsigned long long>	h_valid_count(raw_csv->num_active_cols);
+		CUDA_TRY( cudaMemcpy(h_valid_count.data(), d_valid_count, sizeof(unsigned long long) * h_valid_count.size(), cudaMemcpyDeviceToHost));
 
-		stringColCount++;
+		//--- set the null count
+		for (size_t col = 0; col < h_valid_count.size(); col++) {
+			cols[col]->null_count = raw_csv->num_records - h_valid_count[col];
+		}
 	}
-
-
-	CUDA_TRY( cudaMemcpy(h_valid_count,d_valid_count, sizeof(unsigned long long) * (raw_csv->num_active_cols), cudaMemcpyDeviceToHost));
-
-	//--- set the null count
-	for ( int col = 0; col < raw_csv->num_active_cols; col++) {
-		cols[col]->null_count = raw_csv->num_records - h_valid_count[col];
-	}
-
-	free(h_valid_count); 
 
 	// free up space that is no longer needed
 	if (h_str_cols != NULL)
@@ -742,7 +747,7 @@ gdf_error read_csv(csv_read_arg *args)
 
 	RMM_TRY( RMM_FREE( raw_csv->recStart, 0 ) ); 
 	RMM_TRY( RMM_FREE( raw_csv->d_parseCol, 0 ) ); 
-	CUDA_TRY( cudaFree ( raw_csv->data) );
+	RMM_TRY( RMM_FREE ( raw_csv->data, 0) );
 
 
 	args->data 			= cols;
@@ -884,7 +889,7 @@ gdf_error uploadDataToDevice(const char* h_uncomp_data, size_t h_uncomp_size, ra
 
 	// Exclude the rows user chose to skip at the end of the file
 	if (raw_csv->skipfooter != 0) {
-		raw_csv->num_records = gdf_size_type(max(raw_csv->num_records - raw_csv->skipfooter, 0ul));
+		raw_csv->num_records = gdf_size_type(max(raw_csv->num_records - raw_csv->skipfooter, gdf_size_type{0}));
 		
 	}
 	
@@ -904,11 +909,12 @@ gdf_error uploadDataToDevice(const char* h_uncomp_data, size_t h_uncomp_size, ra
 	// Update the record starts to match the device data (skip missing records, fix offset)
 	for (gdf_size_type i = first_row; i <= first_row + raw_csv->num_records; ++i)
 		h_rec_starts[i] -= start_offset;
+	RMM_TRY(RMM_REALLOC(&raw_csv->recStart, sizeof(cu_recstart_t) * (raw_csv->num_records + 1), 0));
 	CUDA_TRY( cudaMemcpy(raw_csv->recStart, h_rec_starts.data() + first_row, 
 		sizeof(cu_recstart_t) * (raw_csv->num_records + 1), cudaMemcpyDefault));
 
 	// Allocate and copy to the GPU
-	CUDA_TRY(cudaMallocManaged ((void**)&raw_csv->data, (sizeof(char) * raw_csv->num_bytes)));
+	RMM_TRY(RMM_ALLOC ((void**)&raw_csv->data, (sizeof(char) * raw_csv->num_bytes), 0));
 	CUDA_TRY(cudaMemcpy(raw_csv->data, h_uncomp_data + start_offset, raw_csv->num_bytes, cudaMemcpyHostToDevice));
 
 	return GDF_SUCCESS;
@@ -1003,12 +1009,12 @@ gdf_error launch_countRecords(const char* h_data, size_t h_size,
 	vector<cu_reccnt_t> h_cnts(chunk_count);
 
 	cu_reccnt_t* d_cnts = nullptr;
-	CUDA_TRY(cudaMalloc(&d_cnts, sizeof(cu_reccnt_t)* chunk_count));
+	RMM_TRY(RMM_ALLOC (&d_cnts, sizeof(cu_reccnt_t)* chunk_count, 0));
 	CUDA_TRY(cudaMemset(d_cnts, 0, sizeof(cu_reccnt_t)* chunk_count));
 
 	char* d_chunk = nullptr;
 	// Allocate extra byte in case \r\n is at the chunk border
-	CUDA_TRY(cudaMalloc(&d_chunk, max_chunk_bytes + 1)); 
+	RMM_TRY(RMM_ALLOC (&d_chunk, max_chunk_bytes + 1, 0)); 
 
 	int blockSize;		// suggested thread count to use
 	int minGridSize;	// minimum block count required
@@ -1030,8 +1036,8 @@ gdf_error launch_countRecords(const char* h_data, size_t h_size,
 	}
 	CUDA_TRY(cudaMemcpy(h_cnts.data(), d_cnts, chunk_count*sizeof(cu_reccnt_t), cudaMemcpyDefault));
 
-	CUDA_TRY(cudaFree(d_chunk));
-	CUDA_TRY(cudaFree(d_cnts));
+	RMM_TRY( RMM_FREE(d_chunk, 0) );
+	RMM_TRY( RMM_FREE(d_cnts, 0) );
 
 	CUDA_TRY(cudaGetLastError());
 
@@ -1113,7 +1119,7 @@ gdf_error launch_storeRecordStart(const char* h_data, size_t h_size, raw_csv_t *
 
 	char* d_chunk = nullptr;
 	// Allocate extra byte in case \r\n is at the chunk border
-	CUDA_TRY(cudaMalloc(&d_chunk, max_chunk_bytes + 1)); 
+	RMM_TRY(RMM_ALLOC (&d_chunk, max_chunk_bytes + 1, 0)); 
 	
     cu_reccnt_t*	d_num_records;
 	RMM_TRY(RMM_ALLOC((void**)&d_num_records, sizeof(cu_reccnt_t), 0) );
@@ -1146,7 +1152,7 @@ gdf_error launch_storeRecordStart(const char* h_data, size_t h_size, raw_csv_t *
 	}
 
 	RMM_TRY( RMM_FREE( d_num_records, 0 ) ); 
-	CUDA_TRY(cudaFree(d_chunk));
+	RMM_TRY( RMM_FREE( d_chunk, 0 ) );
 
 	CUDA_TRY( cudaGetLastError() );
 
@@ -1333,7 +1339,7 @@ __global__ void convertCsvToGdf(
 
 			// Modify start & end to ignore whitespace and quotechars
 			if(dtype[actual_col] != gdf_dtype::GDF_CATEGORY && dtype[actual_col] != gdf_dtype::GDF_STRING){
-				adjustForWhitespaceAndQuotes(raw_csv, start, tempPos, opts.quotechar);
+				adjustForWhitespaceAndQuotes(raw_csv, &start, &tempPos, opts.quotechar);
 			}
 
 			if(start<=(tempPos)) { // Empty strings are not legal values
@@ -1519,7 +1525,6 @@ __global__ void dataTypeDetection(
 		column_data_t* d_columnData
 		)
 {
-
 	// thread IDs range per block, so also need the block id
 	long	rec_id  = threadIdx.x + (blockDim.x * blockIdx.x);		// this is entry into the field array - tid is an elements within the num_entries array
 
@@ -1589,11 +1594,11 @@ __global__ void dataTypeDetection(
 			long countColon=0;
 			long countString=0;
 
-			long strLen=pos-start;
-
 			// Modify start & end to ignore whitespace and quotechars
 			// This could possibly result in additional empty fields
-			adjustForWhitespaceAndQuotes(raw_csv, start, tempPos);
+			adjustForWhitespaceAndQuotes(raw_csv, &start, &tempPos);
+
+			long strLen=tempPos-start+1;
 
 			for(long startPos=start; startPos<=tempPos; startPos++){
 				if(raw_csv[startPos]>= '0' && raw_csv[startPos] <= '9'){
